+++ tests/test_docker.sh
@@ -12,10 +12,10 @@ mkdir -p run/${DATE}
 READS_R1=/data/AAYHL_HD13M/MG2v3_HD13M_BC13_AGTCAA_L001_R1_001.fastq
 READS_R2=/data/AAYHL_HD13M/MG2v3_HD13M_BC13_AGTCAA_L001_R2_001.fastq
 
-docker run -it -v $DATA_DIR:/data:z $IMAGE preprocess-phix \
+docker run -v $DATA_DIR:/data:z $IMAGE preprocess-phix \
 	-s $READS_R1 -o $OUT_DIR -p $NPROC \
 	| tee run/${DATE}/run_phix_1.out
-docker run -it -v $DATA_DIR:/data:z $IMAGE preprocess-phix \
+docker run -v $DATA_DIR:/data:z $IMAGE preprocess-phix \
 	-s $READS_R2 -o $OUT_DIR -p $NPROC \
 	| tee run/${DATE}/run_phix_2.out
 	
@@ -25,7 +25,7 @@ READS_R2=/data/AAYHL_HD13M/MG2v3_HD13M_BC13_AGTCAA_L001_R2_001.fastq
 YAML=/data/report.yaml
 OUT_DIR="/data/run/${DATE}/presto/${SAMPLE_NAME}"
 
-docker run -it -v $DATA_DIR:/data:z $IMAGE presto-abseq \
+docker run -v $DATA_DIR:/data:z $IMAGE presto-abseq \
     -1 $READS_R1 -2 $READS_R2 -y $YAML -n $SAMPLE_NAME -o $OUT_DIR -p $NPROC \
     | tee run/${DATE}/run_presto.out
 
@@ -33,7 +33,7 @@ docker run -it -v $DATA_DIR:/data:z $IMAGE presto-abseq \
 READS="/data/run/${DATE}/presto/${SAMPLE_NAME}/${SAMPLE_NAME}-final_collapse-unique_atleast-2.fastq"
 OUT_DIR="/data/run/${DATE}/changeo/${SAMPLE_NAME}"
 
-docker run -it -v $DATA_DIR:/data:z $IMAGE changeo-igblast \
+docker run -v $DATA_DIR:/data:z $IMAGE changeo-igblast \
     -s $READS -n $SAMPLE_NAME -o $OUT_DIR -p $NPROC \
     | tee run/${DATE}/run_igblast.out
 
@@ -41,7 +41,7 @@ docker run -it -v $DATA_DIR:/data:z $IMAGE changeo-igblast \
 DB="/data/run/${DATE}/changeo/${SAMPLE_NAME}/${SAMPLE_NAME}_db-pass.tab"
 OUT_DIR="/data/run/${DATE}/changeo/${SAMPLE_NAME}"
 
-docker run -it -v $DATA_DIR:/data:z $IMAGE tigger-genotype \
+docker run -v $DATA_DIR:/data:z $IMAGE tigger-genotype \
     -d $DB -n $SAMPLE_NAME -o $OUT_DIR -p $NPROC \
     | tee run/${DATE}/run_genotype.out
 
@@ -49,7 +49,7 @@ docker run -it -v $DATA_DIR:/data:z $IMAGE tigger-genotype \
 DB="/data/run/${DATE}/changeo/${SAMPLE_NAME}/${SAMPLE_NAME}_genotyped.tab"
 OUT_DIR="/data/run/${DATE}/changeo/${SAMPLE_NAME}"
 
-docker run -it -v $DATA_DIR:/data:z $IMAGE shazam-threshold \
+docker run -v $DATA_DIR:/data:z $IMAGE shazam-threshold \
 	-d $DB -n $SAMPLE_NAME -o $OUT_DIR -p $NPROC \
 	| tee run/${DATE}/run_threshold.out
 
@@ -58,6 +58,6 @@ DB="/data/run/${DATE}/changeo/${SAMPLE_NAME}/${SAMPLE_NAME}_genotyped.tab"
 OUT_DIR="/data/run/${DATE}/changeo/${SAMPLE_NAME}"
 DIST=0.15
 
-docker run -it -v $DATA_DIR:/data:z  $IMAGE changeo-clone \
+docker run -v $DATA_DIR:/data:z $IMAGE changeo-clone \
 	-d $DB -x $DIST -n $SAMPLE_NAME -o $OUT_DIR -p $NPROC \
 	| tee run/${DATE}/run_clone.out